refactor: separated implementations of tasks from task interfaces
